requirements.txt
supabase
notion
openai
python-dotenv